pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompt-eval"
version = "0.1.0"
description = "Prompt evaluation system with FastAPI"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "boto3>=1.34.0",
    "tiktoken>=0.5.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "aiosqlite>=0.19.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "wikipedia>=1.4.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]